add scoper functions and reorder external
