+++ .github/workflows/tests_weekly.yml
@@ -100,6 +100,7 @@ jobs:
             shazam
             alakazam
             tigger
+            scoper
             edgeR
 
       - name: Install Dandelion
