+++ src/dandelion/preprocessing/__init__.py
@@ -1,19 +1,21 @@
 #!/usr/bin/env python
-from dandelion.preprocessing import external
 from dandelion.preprocessing._preprocessing import (
     assign_isotype,
     assign_isotypes,
-    calculate_threshold,
     check_contigs,
     create_germlines,
     filter_contigs,
     format_fasta,
     format_fastas,
-    quantify_mutations,
     reannotate_genes,
     reassign_alleles,
 )
-from dandelion.preprocessing.external._preprocessing import recipe_scanpy_qc
+from dandelion.external.immcantation.shazam import (
+    calculate_threshold,
+    quantify_mutations,
+)
+
+from dandelion.external.scanpy import recipe_scanpy_qc
 
 __all__ = [
     "assign_isotype",
@@ -21,7 +23,6 @@ __all__ = [
     "calculate_threshold",
     "check_contigs",
     "create_germlines",
-    "external",
     "filter_contigs",
     "format_fasta",
     "format_fastas",
