--- src/dandelion/preprocessing/external/__init__.py
@@ -1,20 +1,0 @@
-#!/usr/bin/env python
-from ._preprocessing import (
-    assigngenes_igblast,
-    creategermlines,
-    makedb_igblast,
-    parsedb_heavy,
-    parsedb_light,
-    recipe_scanpy_qc,
-    tigger_genotype,
-)
-
-__all__ = [
-    "assigngenes_igblast",
-    "creategermlines",
-    "makedb_igblast",
-    "parsedb_heavy",
-    "parsedb_light",
-    "recipe_scanpy_qc",
-    "tigger_genotype",
-]
