+++ tests/test_rpy2.py
@@ -1,11 +1,12 @@
 #!/usr/bin/env python
-import pandas as pd
-import dandelion as ddl
-import sys
 import pytest
 
+# import sys
+import pandas as pd
+import dandelion as ddl
 
-@pytest.mark.skipif(sys.platform == "darwin", reason="macos CI stalls.")
+
+# @pytest.mark.skipif(sys.platform == "darwin", reason="macos CI stalls.")
 def test_importrpy2():
     """test_importrpy2"""
 
@@ -17,7 +18,7 @@ def test_importrpy2():
 
 
 @pytest.mark.usefixtures("create_testfolder", "airr_reannotated")
-@pytest.mark.skipif(sys.platform == "darwin", reason="macos CI stalls.")
+# @pytest.mark.skipif(sys.platform == "darwin", reason="macos CI stalls.")
 def test_mutation(create_testfolder, airr_reannotated):
     """test_mutation"""
     f = create_testfolder / "test.tsv"
@@ -31,7 +32,7 @@ def test_mutation(create_testfolder, airr_reannotated):
 
 
 @pytest.mark.usefixtures("create_testfolder", "database_paths")
-@pytest.mark.skipif(sys.platform == "darwin", reason="macos CI stalls.")
+# @pytest.mark.skipif(sys.platform == "darwin", reason="macos CI stalls.")
 def test_create_germlines(create_testfolder, database_paths):
     """test create germlines"""
     f = create_testfolder / "test.tsv"
@@ -42,7 +43,7 @@ def test_create_germlines(create_testfolder, database_paths):
 
 
 @pytest.mark.usefixtures("create_testfolder")
-@pytest.mark.skipif(sys.platform == "darwin", reason="macos CI stalls.")
+# @pytest.mark.skipif(sys.platform == "darwin", reason="macos CI stalls.")
 def test_manual_threshold_and_define_clones(create_testfolder):
     """test threshold"""
     f = create_testfolder / "test.tsv"
@@ -56,7 +57,7 @@ def test_manual_threshold_and_define_clones(create_testfolder):
 
 
 @pytest.mark.usefixtures("create_testfolder")
-@pytest.mark.skipif(sys.platform == "darwin", reason="macos CI stalls.")
+# @pytest.mark.skipif(sys.platform == "darwin", reason="macos CI stalls.")
 def test_define_clones_outdir(create_testfolder):
     """test threshold"""
     f = create_testfolder / "test.tsv"
@@ -69,3 +70,52 @@ def test_define_clones_outdir(create_testfolder):
     ddl.tl.define_clones(vdj, out_dir=out_path)
     assert len(list(out_path.iterdir())) == 3
     assert len(list((out_path / "tmp").iterdir())) == 2
+
+
+@pytest.mark.usefixtures("create_testfolder")
+# @pytest.mark.skipif(sys.platform == "darwin", reason="macos CI stalls.")
+def test_scoper_i(create_testfolder):
+    """test identical clones from scoper"""
+    f = create_testfolder / "test.tsv"
+    vdj = ddl.Dandelion(f)
+    assert "clone_id" not in vdj.data
+    from dandelion.external.immcantation.scoper import identical_clones
+
+    identical_clones(vdj)
+    assert not vdj.data.clone_id.empty
+
+
+@pytest.mark.usefixtures("create_testfolder")
+# @pytest.mark.skipif(sys.platform == "darwin", reason="macos CI stalls.")
+def test_scoper_h(create_testfolder):
+    """test hierarchical clones from scoper"""
+    f = create_testfolder / "test.tsv"
+    vdj = ddl.Dandelion(f)
+    assert "clone_id" not in vdj.data
+    from dandelion.external.immcantation.scoper import hierarchical_clones
+
+    hierarchical_clones(vdj, threshold=0.15)
+    assert not vdj.data.clone_id.empty
+
+
+@pytest.mark.usefixtures("create_testfolder")
+# @pytest.mark.skipif(sys.platform == "darwin", reason="macos CI stalls.")
+def test_scoper_h(create_testfolder):
+    """test spectral clones from scoper"""
+    f = create_testfolder / "test.tsv"
+    vdj = ddl.Dandelion(f)
+    assert "clone_id" not in vdj.data
+    from dandelion.external.immcantation.scoper import spectral_clones
+
+    spectral_clones(vdj, method="novj")
+    assert not vdj.data.clone_id.empty
+
+    vdj = ddl.Dandelion(f)
+    assert "clone_id" not in vdj.data
+    spectral_clones(vdj, method="novj", threshold=0.15)
+    assert not vdj.data.clone_id.empty
+
+    vdj = ddl.Dandelion(f)
+    assert "clone_id" not in vdj.data
+    spectral_clones(vdj, method="vj", threshold=0.15)
+    assert not vdj.data.clone_id.empty
